requirements.txt
flask
twilio
python-dotenv
opencv-python
ultralytics
face_recognition
numpy